requirements.txt
pandas
plotly.express
streamlit
sklearn
scikit-learn
datetime
